updated migrations

+++ database/migrations/2019_01_29_225555_create_products_table.php
@@ -14,8 +14,8 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('category_id');
+            $table->bigIncrements('id');
+            $table->unsignedInteger('category_id');
             $table->string('title');
             $table->string('slug');
             $table->string('mpn', 50)->nullable();  // Manufacturer Part Number
@@ -26,7 +26,7 @@ class CreateProductsTable extends Migration
             $table->text('description')->nullable();
             $table->unsignedDecimal('min_price', 9, 2);
             $table->unsignedDecimal('max_price', 9, 2);
-            $table->integer('brand_id')->nullable();
+            $table->unsignedInteger('brand_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
